is_identifier

+++ test/StringIO_Test.cpp
@@ -70,7 +70,7 @@ TEST(StringIO_Test, random)
 
  for (int j = N; --j >= 0;)
  {
-  for (int i = 0; i < Length; i++)
+  for (size_t i = 0; i < Length; i++)
    s1[i] = char(ud(mt));
 
   std::ostringstream out;
